added api and added events

+++ src/main/java/de/blocki/advancedscoreboard/listener/PlayerJoinLeave.java
@@ -1,8 +1,10 @@
 package de.blocki.advancedscoreboard.listener;
 
+import de.blocki.advancedscoreboard.api.events.FBScoreboardRemoveEvent;
+import de.blocki.advancedscoreboard.api.events.FBScoreboardSetEvent;
 import de.blocki.advancedscoreboard.fb.FastBoard;
-import de.blocki.advancedscoreboard.main.Main;
-import de.blocki.advancedscoreboard.utils.ConfigManager;
+import de.blocki.advancedscoreboard.main.AdvancedScoreboard;
+import org.bukkit.Bukkit;
 import org.bukkit.entity.Player;
 import org.bukkit.event.EventHandler;
 import org.bukkit.event.Listener;
@@ -12,23 +14,41 @@ import org.bukkit.event.player.PlayerQuitEvent;
 public class PlayerJoinLeave implements Listener {
 
     @EventHandler
-    public void onPlayerJoin(final PlayerJoinEvent event) {
+    public void onPlayerJoin(PlayerJoinEvent event) {
+        //gets Player from event
         Player player = event.getPlayer();
 
+        //creates new FastBoard
         FastBoard board = new FastBoard(player);
 
-        board.updateTitle(Main.fbScoreboard.getTitle().replace("&", "§"));
+        //call the FBScoreboardSetEvent
+        FBScoreboardSetEvent e = new FBScoreboardSetEvent(player, board);
+        Bukkit.getServer().getPluginManager().callEvent(e);
 
-        Main.boards.put(player.getUniqueId(), board);
+        if(!e.isCancelled()) {
+
+            //set the config title
+            board.updateTitle(AdvancedScoreboard.fbScoreboard.getTitle().replace("&", "§"));
+
+            //sets the board in main cache
+            AdvancedScoreboard.boards.put(player.getUniqueId(), board);
+
+        }
+
+
     }
 
     @EventHandler
-    public void onQuit(PlayerQuitEvent e) {
-        Player player = e.getPlayer();
+    public void onQuit(PlayerQuitEvent event) {
+        Player player = event.getPlayer();
 
-        FastBoard board = Main.boards.remove(player.getUniqueId());
+        FastBoard board = AdvancedScoreboard.boards.remove(player.getUniqueId());
 
-        if (board != null) {
+        //call the FBScoreboardRemoveEvent
+        FBScoreboardRemoveEvent e = new FBScoreboardRemoveEvent(player, board);
+        Bukkit.getServer().getPluginManager().callEvent(e);
+
+        if (board != null && !e.isCancelled()) {
             board.delete();
         }
     }
